Github_API/Harvester/requirement.txt
cloudant
nltk
PyGithub
pydriller